split file and modify name
