+++ src/parse_part.c
@@ -198,11 +198,6 @@ Node *call_function(Node *node, Token *tok){
 	return node;
 }
 
-TypeKind get_pointer_type(Type *given){
-	while(given->ptr_to != NULL) given = given->ptr_to;
-	return given->ty;
-}
-
 Node *array_index(Node *node, Node *index){
 	Node *pointer_size;
 
@@ -215,28 +210,6 @@ Node *array_index(Node *node, Node *index){
 	node->rhs = new_node(ND_MUL, index, pointer_size);
 
 	node = new_node(ND_DEREF, NULL, node);
-
-	return node;
-}
-
-Node *pointer_calc(Node *node, Type lhs_type, Type rhs_type){
-	int ptrtype;
-
-
-	node->type.ty = PTR;
-	Node *pointer_size = calloc(1, sizeof(Node));
-	pointer_size->kind = ND_NUM;
-
-
-	if(type_size(lhs_type.ty) == 8 && lhs_type.ptr_to!=NULL){
-		ptrtype = lhs_type.ptr_to->ty;
-		pointer_size->val = type_size(ptrtype);
-		node->rhs = new_node(ND_MUL, node->rhs, pointer_size);
-	}else if(type_size(rhs_type.ty) == 8 && rhs_type.ptr_to!=NULL){
-		ptrtype = rhs_type.ptr_to->ty;
-		pointer_size->val = type_size(ptrtype);
-		node->lhs = new_node(ND_MUL, node->lhs, pointer_size);
-	}
 
 	return node;
 }
@@ -393,7 +366,7 @@ Node *declare_local_variable(Node *node, Token *tok, int star_count){
 	return node;
 }
 
-int align_array_size(int isize, TypeKind array_type){
-	int array_size = isize*type_size(array_type);
-	return (array_size%8) ? array_size/8*8+8 : array_size;
+
+void declare_struct(){
+	return;
 }
